refactor: using index in components folder
