+++ native-client/src/components/Recorder.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import {Button, View, Text, AsyncStorage as store } from 'react-native'
 import Expo, { Audio, Permissions } from 'expo'
 import Timer  from './Timer'
-import {Uploader, RecordButton} from './'
+import { Uploader, RecordButton } from './'
 import styles from '../../assets/stylesheet'
 
 export default class Recorder extends Component {
@@ -108,7 +108,7 @@ export default class Recorder extends Component {
     requestAnimationFrame(this.frame)
   }
 
-  frame = (time) => {
+  frame = () => {
       if (this.state.isClicked) {
         const duration = new Date().getTime() - this.startTime
         this.setState({duration})
@@ -139,7 +139,7 @@ export default class Recorder extends Component {
           <View style={styles.recorderBottomText}>
             <Text style={styles.recorderIntroText} > Press the mic and start speaking </Text>
             <Text style={styles.recorderText}> We will analyze your speech and </Text>
-            <Text style={styles.recorderText}> provide you with suggestions how to improve </Text>
+            <Text style={styles.recorderText}> provide you with suggestions on how to improve </Text>
           </View>
           }
           {(this.state.begin && this.state.isClicked) ?
@@ -148,7 +148,7 @@ export default class Recorder extends Component {
           </View>
           :
           <View style={styles.recorderBottomText}>
-          <Text style={styles.recorderStopText} > </Text>
+          <Text style={styles.recorderStopText} />
           </View>
           }
           <View style={styles.recorderBottomText}>
