+++ native-client/src/components/index.js
@@ -7,3 +7,5 @@ export {default as Logout} from './Logout'
 export {default as Speeches} from './Speeches'
 export {default as RecordButton} from './RecordButton'
 export {default as TimerRender} from './TimerRender'
+export {default as EditModal} from './EditModal'
+export {default as EditTitleForm} from './EditTitleForm'
